requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
python-dotenv==0.20.0
python-http-client==3.3.7
sendgrid==6.9.7
sniffio==1.2.0
starkbank-ecdsa==2.0.3
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
